+++ packages/main/test/sap/ui/webcomponents/main/pages/Simple.html
@@ -18,7 +18,7 @@
 
 <body>
 
-<ui5-datepicker></ui5-datepicker>
+<ui5-button>hi</ui5-button>
 
 </body>
 </html>
